mobile-appium-tests/requirements.txt
Appium-Python-Client==2.11.1
pytest==8.2.2
pytest-junitxml==1.5.0
